WIP, gjør om Behandling til ikke å være sealed class
