+++ domene/src/main/kotlin/no/nav/tiltakspenger/saksbehandling/domene/behandling/Behandling.kt
@@ -4,8 +4,8 @@ import no.nav.tiltakspenger.felles.BehandlingId
 import no.nav.tiltakspenger.felles.Periode
 import no.nav.tiltakspenger.felles.SakId
 import no.nav.tiltakspenger.felles.Saksbehandler
-import no.nav.tiltakspenger.saksbehandling.domene.saksopplysning.Kilde
 import no.nav.tiltakspenger.saksbehandling.domene.saksopplysning.Saksopplysning
+import no.nav.tiltakspenger.saksbehandling.domene.vilkår.Vurdering
 
 data class LeggTilSaksopplysningRespons(
     val behandling: Behandling,
@@ -16,42 +16,24 @@ interface Behandling {
     val id: BehandlingId
     val sakId: SakId
     val vurderingsperiode: Periode
+    val søknader: List<Søknad>
+    val saksbehandler: String?
+    val beslutter: String?
     val saksopplysninger: List<Saksopplysning>
     val tiltak: List<Tiltak>
-    val saksbehandler: String?
+    val vilkårsvurderinger: List<Vurdering>
     val utfallsperioder: List<Utfallsperiode>
-    val søknader: List<Søknad>
+    val status: BehandlingStatus
+    val tilstand: BehandlingTilstand
 
-    fun søknad(): Søknad = sisteSøknadMedOpprettetFraFørste()
-
-    private fun sisteSøknadMedOpprettetFraFørste(): Søknad =
-        søknader.maxBy { it.opprettet }.copy(opprettet = søknader.minBy { it.opprettet }.opprettet)
-
-    fun saksopplysninger(): List<Saksopplysning> {
-        return saksopplysninger.groupBy { it.vilkår }.map { entry ->
-            entry.value.reduce { acc, saksopplysning ->
-                if (saksopplysning.kilde == Kilde.SAKSB) saksopplysning else acc
-            }
-        }
-    }
-
-    fun leggTilSøknad(søknad: Søknad): BehandlingVilkårsvurdert {
-        throw IllegalStateException("Kan ikke legge til søknad på denne behandlingen")
-    }
-
-    fun leggTilSaksopplysning(saksopplysning: Saksopplysning): LeggTilSaksopplysningRespons {
-        throw IllegalStateException("Kan ikke legge til saksopplysning på denne behandlingen")
-    }
-
-    fun oppdaterTiltak(tiltak: List<Tiltak>): Behandling {
-        throw IllegalStateException("Kan ikke oppdatere tiltak på denne behandlingen")
-    }
-
-    fun startBehandling(saksbehandler: Saksbehandler): Behandling {
-        throw IllegalStateException("Kan ikke starte en behandling med denne statusen")
-    }
-
-    fun avbrytBehandling(saksbehandler: Saksbehandler): Behandling {
-        throw IllegalStateException("Kan ikke avbryte en behandling med denne statusen")
-    }
+    fun leggTilSøknad(søknad: Søknad): Behandling
+    fun leggTilSaksopplysning(saksopplysning: Saksopplysning): LeggTilSaksopplysningRespons
+    fun oppdaterTiltak(tiltak: List<Tiltak>): Behandling
+    fun startBehandling(saksbehandler: Saksbehandler): Behandling
+    fun avbrytBehandling(saksbehandler: Saksbehandler): Behandling
+    fun tilBeslutting(saksbehandler: Saksbehandler): Behandling
+    fun iverksett(utøvendeBeslutter: Saksbehandler): Behandling
+    fun sendTilbake(utøvendeBeslutter: Saksbehandler): Behandling
+    fun vilkårsvurder(): Behandling
+    fun saksopplysninger(): List<Saksopplysning>
 }
