--- domene/src/main/kotlin/no/nav/tiltakspenger/saksbehandling/domene/behandling/BehandlingOpprettet.kt
@@ -1,86 +1,0 @@
-package no.nav.tiltakspenger.saksbehandling.domene.behandling
-
-import no.nav.tiltakspenger.felles.BehandlingId
-import no.nav.tiltakspenger.felles.Periode
-import no.nav.tiltakspenger.felles.SakId
-import no.nav.tiltakspenger.felles.Saksbehandler
-import no.nav.tiltakspenger.saksbehandling.domene.saksopplysning.Saksopplysning
-import no.nav.tiltakspenger.saksbehandling.domene.saksopplysning.Saksopplysninger
-import no.nav.tiltakspenger.saksbehandling.domene.saksopplysning.Saksopplysninger.oppdaterSaksopplysninger
-import no.nav.tiltakspenger.saksbehandling.domene.vilkår.vilkårsvurder
-
-data class BehandlingOpprettet(
-    override val id: BehandlingId,
-    override val sakId: SakId,
-    override val søknader: List<Søknad>,
-    override val vurderingsperiode: Periode,
-    override val saksopplysninger: List<Saksopplysning>,
-    override val tiltak: List<Tiltak>,
-    override val saksbehandler: String?,
-    override val utfallsperioder: List<Utfallsperiode> = emptyList(),
-) : Førstegangsbehandling {
-
-    companion object {
-
-        fun opprettBehandling(sakId: SakId, søknad: Søknad): BehandlingOpprettet {
-            return BehandlingOpprettet(
-                id = BehandlingId.random(),
-                sakId = sakId,
-                søknader = listOf(søknad),
-                vurderingsperiode = søknad.vurderingsperiode(),
-                saksopplysninger = Saksopplysninger.initSaksopplysningerFraSøknad(søknad) + Saksopplysninger.lagSaksopplysningerAvSøknad(
-                    søknad,
-                ),
-                tiltak = emptyList(),
-                saksbehandler = null,
-            )
-        }
-    }
-
-    override fun leggTilSøknad(søknad: Søknad): BehandlingVilkårsvurdert {
-        val fakta = if (søknad.vurderingsperiode() != this.vurderingsperiode) {
-            Saksopplysninger.initSaksopplysningerFraSøknad(søknad) +
-                Saksopplysninger.lagSaksopplysningerAvSøknad(søknad)
-        } else {
-            Saksopplysninger.lagSaksopplysningerAvSøknad(søknad)
-                .fold(this.saksopplysninger) { acc, saksopplysning ->
-                    acc.oppdaterSaksopplysninger(saksopplysning)
-                }
-        }
-
-        return this.copy(
-            søknader = this.søknader + søknad,
-            vurderingsperiode = søknad.vurderingsperiode(),
-            saksopplysninger = fakta,
-        ).vilkårsvurder()
-    }
-
-    override fun leggTilSaksopplysning(saksopplysning: Saksopplysning): LeggTilSaksopplysningRespons {
-        val oppdatertSaksopplysningListe = saksopplysninger.oppdaterSaksopplysninger(saksopplysning)
-        return if (oppdatertSaksopplysningListe == this.saksopplysninger) {
-            LeggTilSaksopplysningRespons(
-                behandling = this,
-                erEndret = false,
-            )
-        } else {
-            LeggTilSaksopplysningRespons(
-                behandling = this.copy(saksopplysninger = oppdatertSaksopplysningListe).vilkårsvurder(),
-                erEndret = true,
-            )
-        }
-    }
-
-    override fun oppdaterTiltak(tiltak: List<Tiltak>): Førstegangsbehandling =
-        this.copy(tiltak = tiltak).vilkårsvurder()
-
-    override fun startBehandling(saksbehandler: Saksbehandler): Førstegangsbehandling {
-        check(this.saksbehandler == null) { "Denne behandlingen er allerede tatt" }
-        check(saksbehandler.isSaksbehandler()) { "Saksbehandler må være saksbehandler" }
-        return this.copy(saksbehandler = saksbehandler.navIdent)
-    }
-
-    override fun avbrytBehandling(saksbehandler: Saksbehandler): Førstegangsbehandling {
-        check(saksbehandler.isSaksbehandler() || saksbehandler.isAdmin()) { "Kan ikke avbryte en behandling som ikke er din" }
-        return this.copy(saksbehandler = null)
-    }
-}
