+++ domene/src/main/kotlin/no/nav/tiltakspenger/saksbehandling/domene/behandling/BehandlingTilstand.kt
@@ -1,0 +1,5 @@
+package no.nav.tiltakspenger.saksbehandling.domene.behandling
+
+enum class BehandlingTilstand {
+    OPPRETTET, VILKÅRSVURDERT, TIL_BESLUTTER, IVERKSATT
+}
