+++ domene/src/main/kotlin/no/nav/tiltakspenger/saksbehandling/domene/vedtak/Vedtak.kt
@@ -3,7 +3,7 @@ package no.nav.tiltakspenger.saksbehandling.domene.vedtak
 import no.nav.tiltakspenger.felles.Periode
 import no.nav.tiltakspenger.felles.SakId
 import no.nav.tiltakspenger.felles.VedtakId
-import no.nav.tiltakspenger.saksbehandling.domene.behandling.BehandlingIverksatt
+import no.nav.tiltakspenger.saksbehandling.domene.behandling.Behandling
 import no.nav.tiltakspenger.saksbehandling.domene.behandling.Utfallsperiode
 import no.nav.tiltakspenger.saksbehandling.domene.saksopplysning.Saksopplysning
 import no.nav.tiltakspenger.saksbehandling.domene.vilkår.Vurdering
@@ -12,7 +12,7 @@ import java.time.LocalDateTime
 data class Vedtak(
     val id: VedtakId = VedtakId.random(),
     val sakId: SakId,
-    val behandling: BehandlingIverksatt,
+    val behandling: Behandling,
     val vedtaksdato: LocalDateTime,
     val vedtaksType: VedtaksType,
     val periode: Periode,
