+++ domene/src/main/kotlin/no/nav/tiltakspenger/saksbehandling/ports/MultiRepo.kt
@@ -1,17 +1,16 @@
 package no.nav.tiltakspenger.saksbehandling.ports
 
 import no.nav.tiltakspenger.saksbehandling.domene.attestering.Attestering
-import no.nav.tiltakspenger.saksbehandling.domene.behandling.BehandlingIverksatt
-import no.nav.tiltakspenger.saksbehandling.domene.behandling.BehandlingVilkårsvurdert
+import no.nav.tiltakspenger.saksbehandling.domene.behandling.Behandling
 import no.nav.tiltakspenger.saksbehandling.domene.vedtak.Vedtak
 
 interface MultiRepo {
     suspend fun lagreOgKjør(
-        iverksattBehandling: BehandlingIverksatt,
+        iverksattBehandling: Behandling,
         attestering: Attestering,
         vedtak: Vedtak,
         operasjon: suspend () -> String,
     ): String
 
-    fun lagre(behandling: BehandlingVilkårsvurdert, attestering: Attestering)
+    fun lagre(behandling: Behandling, attestering: Attestering)
 }
