+++ domene/src/main/kotlin/no/nav/tiltakspenger/saksbehandling/service/behandling/BehandlingServiceImpl.kt
@@ -10,12 +10,9 @@ import no.nav.tiltakspenger.felles.exceptions.IkkeFunnetException
 import no.nav.tiltakspenger.saksbehandling.domene.attestering.Attestering
 import no.nav.tiltakspenger.saksbehandling.domene.attestering.AttesteringStatus
 import no.nav.tiltakspenger.saksbehandling.domene.behandling.Behandling
-import no.nav.tiltakspenger.saksbehandling.domene.behandling.BehandlingIverksatt
 import no.nav.tiltakspenger.saksbehandling.domene.behandling.BehandlingStatus
-import no.nav.tiltakspenger.saksbehandling.domene.behandling.BehandlingTilBeslutter
-import no.nav.tiltakspenger.saksbehandling.domene.behandling.BehandlingVilkårsvurdert
+import no.nav.tiltakspenger.saksbehandling.domene.behandling.BehandlingTilstand
 import no.nav.tiltakspenger.saksbehandling.domene.behandling.Førstegangsbehandling
-import no.nav.tiltakspenger.saksbehandling.domene.behandling.RevurderingOpprettet
 import no.nav.tiltakspenger.saksbehandling.domene.behandling.Revurderingsbehandling
 import no.nav.tiltakspenger.saksbehandling.domene.behandling.Tiltak
 import no.nav.tiltakspenger.saksbehandling.domene.saksopplysning.Saksopplysning
@@ -85,7 +82,7 @@ class BehandlingServiceImpl(
     ) {
         check(utøvendeSaksbehandler.roller.contains(Rolle.SAKSBEHANDLER)) { "Saksbehandler må være saksbehandler" }
         val behandling = hentBehandling(behandlingId)
-        if (behandling is BehandlingVilkårsvurdert) {
+        if (behandling.tilstand == BehandlingTilstand.VILKÅRSVURDERT) {
             behandlingRepo.lagre(behandling.tilBeslutting(utøvendeSaksbehandler))
         }
     }
@@ -105,8 +102,8 @@ class BehandlingServiceImpl(
             beslutter = utøvendeBeslutter.navIdent,
         )
 
-        when (behandling) {
-            is BehandlingTilBeslutter -> {
+        when (behandling.tilstand) {
+            BehandlingTilstand.TIL_BESLUTTER -> {
                 multiRepo.lagre(behandling.sendTilbake(utøvendeBeslutter), attestering)
             }
 
@@ -119,8 +116,8 @@ class BehandlingServiceImpl(
         val sak = sakRepo.hentSakDetaljer(behandling.sakId)
             ?: throw IllegalStateException("iverksett finner ikke sak ${behandling.sakId}")
 
-        val iverksattBehandling = when (behandling) {
-            is BehandlingTilBeslutter -> behandling.iverksett(utøvendeBeslutter)
+        val iverksattBehandling = when (behandling.tilstand) {
+            BehandlingTilstand.TIL_BESLUTTER -> behandling.iverksett(utøvendeBeslutter)
             else -> throw IllegalStateException("Behandlingen har feil tilstand og kan ikke iverksettes. BehandlingId: $behandlingId")
         }
         val attestering = Attestering(
@@ -143,7 +140,10 @@ class BehandlingServiceImpl(
         brevPublisherGateway.sendBrev(sak.saknummer, vedtak, personopplysninger)
     }
 
-    private fun lagVedtakForBehandling(behandling: BehandlingIverksatt): Vedtak {
+    private fun lagVedtakForBehandling(behandling: Behandling): Vedtak {
+        require(behandling.tilstand == BehandlingTilstand.IVERKSATT) { "Kan ikke lage vedtakk for behandling som ikke er iverksatt" }
+        require(behandling.saksbehandler != null) { "Kan ikke lage vedtakk for behandling som ikke har saksbehandler" }
+        require(behandling.beslutter != null) { "Kan ikke lage vedtakk for behandling som ikke har beslutter" }
         return Vedtak(
             id = VedtakId.random(),
             sakId = behandling.sakId,
@@ -154,8 +154,8 @@ class BehandlingServiceImpl(
             periode = behandling.vurderingsperiode,
             saksopplysninger = behandling.saksopplysninger(),
             vurderinger = behandling.vilkårsvurderinger,
-            saksbehandler = behandling.saksbehandler,
-            beslutter = behandling.beslutter,
+            saksbehandler = behandling.saksbehandler!!,
+            beslutter = behandling.beslutter!!,
         )
     }
 
@@ -185,7 +185,7 @@ class BehandlingServiceImpl(
         check(utøvendeSaksbehandler.roller.contains(Rolle.SAKSBEHANDLER)) { "Saksbehandler må være saksbehandler" }
 
         val vedtak = vedtakRepo.hentVedtakForBehandling(behandlingId)
-        val revurderingBehandling = RevurderingOpprettet.opprettRevurderingsbehandling(
+        val revurderingBehandling = Revurderingsbehandling.opprettRevurderingsbehandling(
             vedtak = vedtak,
             navIdent = utøvendeSaksbehandler.navIdent,
         )
